[Update]  convert to Eupi mode

+++ app/src/main/java/com/euphonyio/orderwith/RequestCodeEnum.kt
@@ -1,6 +1,6 @@
 package com.euphonyio.orderwith
 
-enum class RequestCodeEnum(val code: Int) {
-    MENU_REQUEST(18500),
-    ORDER_REQUEST(20500)
+enum class RequestCodeEnum(val code: Double) {
+    MENU_REQUEST(18500.0),
+    ORDER_REQUEST(20500.0)
 }
